11.06

+++ utils/scrappdf.py
@@ -8,7 +8,8 @@ from scrapy.http import Request
 
 
 class PdfSearch(scrapy.Spider):
-    name = 'pdf_search'
+
+    name = 'FileScrap'
 
     custom_settings = {
         'HTTPERROR_ALLOW_ALL': True,
@@ -17,29 +18,44 @@ class PdfSearch(scrapy.Spider):
         'RETRY_HTTP_CODES': [],
     }
 
-    def __init__(self, *args, **kwargs):
-        super(PdfSearch, self).__init__(*args, **kwargs)
-        self.start_urls = str(kwargs.get('input_url'))
-        self.allowed_domains = [str(kwargs.get('allowed_domains'))]
-
     def start_requests(self):
         logging.getLogger('scrapy').propagate = False
-        url = getattr(self, 'url', self.start_urls)
-        yield scrapy.Request(url, dont_filter=True, callback=self.parse, errback=self.errback)
+        requests = []
+        urlfile = '/home/camelo/PycharmProjects/HackPy-EFEj2/projecte/' + nameproject + '/url.txt'
+        for url in open(urlfile):
+            requests.append(scrapy.Request(url=url, callback=self.parse))
+        return requests
 
     def parse(self, response):
         for href in response.css('a::attr(href)').extract():
-            yield Request(url=response.urljoin(href), callback=self.parse_pdf)
+            yield Request(url=response.urljoin(href), callback=self.parse_file)
 
-    def parse_pdf(self, response):
+    def parse_file(self, response):
         for href in response.css('a[href$=".pdf"]::attr(href)').extract():
-            yield Request(url=response.urljoin(href), callback=self.save_pdf)
+            yield Request(url=response.urljoin(href), callback=self.save)
+        for href in response.css('a[href$=".jpg"]::attr(href)').extract():
+            yield Request(url=response.urljoin(href), callback=self.save)
+        for href in response.css('a[href$=".png"]::attr(href)').extract():
+            yield Request(url=response.urljoin(href), callback=self.save)
+        for href in response.css('img').xpath('@src').extract():
+            yield Request(url=response.urljoin(href), callback=self.save)
 
-    def save_pdf(self, response):
-        f = response.url.split('/')[-1]
-        self.logger.info('Saving PDF %s', f)
-        with open(f, 'wb') as f:
-            f.write(response.body)
+    def save(self, response):
+        import os
+        f = ''
+        path_download_foto = '/home/camelo/PycharmProjects/HackPy-EFEj2/photo_downloaded/'
+        path_download_pdf = '/home/camelo/PycharmProjects/HackPy-EFEj2/pdf_downloaded/'
+        name_file = response.url.split('/')[-1]
+        if name_file.endswith('.jpg') or name_file.endswith('.png'):
+            f = path_download_foto + name_file
+        elif name_file.endswith('.pdf'):
+            f = path_download_pdf + name_file
+        if not os.path.isfile(f) or f == '':
+            self.logger.info('Saving File %s', f)
+            with open(f, 'wb') as f:
+                f.write(response.body)
+        else:
+            self.logger.info('File %s NOT SAVED!!', f)
 
     def errback(self, err):
         """Handles an error"""
